A more succinct statement of the License
The list item stating that the code is "open source" does not mention under which license is applied, whilst this is stated via a link to the repository license below. It would perhaps be more succinct to state directly in the open source statement which license the repository is released under. This removes the need for the subsequent statement, since the rights of the user are described in the license itself.